requirements.txt
easyocr==1.7.1
fastapi==0.111.0
openai==1.35.10
opencv_python==4.7.0.72
opencv-contrib-python
python-dotenv==1.0.1
streamlit
python-dotenv
pypdf2
python-docx
streamlit-extras
langchain
openai
moviepy 
pytube
#brew install ffmpeg
google-generativeai
moviepy
transformers
onnxruntime
torchaudio
matplotlib
whisper-timestamped